Move argparser and example config into Django app
